+++ src/PortentaH7/PortentaH7_WiFiMulti_Generic_Impl.h
@@ -1,5 +1,5 @@
 /**************************************************************************************************************************************
-  NINA_WiFiMulti_Generic_Impl.h
+  PortentaH7_WiFiMulti_Generic_Impl.h
   For any WiFi shields, such as ESP32, ESP8266, Portenta_H7, WiFiNINA W101, W102, W13x, or custom, such as ESP8266/ESP32-AT, etc
   
   WiFiMulti_Generic is a library to adapt the  ESP32/ESP8266 WiFiMulti feature to other WiFi modules
@@ -34,8 +34,8 @@
 
 #pragma once
 
-#ifndef _NINA_WIFIMULTI_GENERIC_IMPL_H_
-#define _NINA_WIFIMULTI_GENERIC_IMPL_H_ 
+#ifndef _PORTENTAH7_WIFIMULTI_GENERIC_IMPL_H_
+#define _PORTENTAH7_WIFIMULTI_GENERIC_IMPL_H_ 
 
 #include <limits.h>
 #include <string.h>
@@ -148,7 +148,6 @@ uint8_t WiFiMulti_Generic::run(const uint32_t& connectTimeout)
     
     int bestNetworkDb = INT_MIN;
     uint8_t bestBSSID[6];
-    int32_t bestChannel = 0;
 
     WFM_LOGINFO("[run] Scan done");
 
@@ -161,16 +160,14 @@ uint8_t WiFiMulti_Generic::run(const uint32_t& connectTimeout)
       WFM_LOGINFO1("[run] Number of Networks found:", scanResult);
       
       for (int8_t i = 0; i < scanResult; ++i) 
-      {       
+      {        
         String ssid_scan    = WiFi.SSID(i);
         int32_t rssi_scan   = WiFi.RSSI(i); 
         uint8_t sec_scan    = WiFi.encryptionType(i); 
         
         byte bssid[6];
-        uint8_t* BSSID_scan  = WiFi.BSSID(i, bssid);
+        uint8_t* BSSID_scan  = WiFi.BSSID(bssid);
         
-        int32_t chan_scan   = WiFi.channel(i);     
-
         bool known = false;
         
         for (uint32_t x = APlist.size() ; x > 0; x--) 
@@ -189,7 +186,6 @@ uint8_t WiFiMulti_Generic::run(const uint32_t& connectTimeout)
               { 
                 // check for passphrase if not open wlan
                 bestNetworkDb = rssi_scan;
-                bestChannel = chan_scan;
                 memcpy((void*) &bestNetwork, (void*) &entry, sizeof(bestNetwork));
                 memcpy((void*) &bestBSSID, (void*) BSSID_scan, sizeof(bestBSSID));
               }
@@ -203,13 +199,13 @@ uint8_t WiFiMulti_Generic::run(const uint32_t& connectTimeout)
 
         if (known) 
         {
-          WFM_LOGDEBUG3("  Known => #", i, ", Channel:", chan_scan);
+          WFM_LOGDEBUG1("  Known => #", i);
           WFM_HEXLOGDEBUG5_SEPARATOR(":", BSSID_scan[0], BSSID_scan[1], BSSID_scan[2], BSSID_scan[3], BSSID_scan[4], BSSID_scan[5]);
           WFM_LOGDEBUG5("SSID:", ssid_scan.c_str(), ", RSSI:", rssi_scan, ", Secured:", (sec_scan == ENC_TYPE_NONE /*WIFI_AUTH_OPEN*/) ? 'n' : 'y');
         } 
         else 
         {
-          WFM_LOGDEBUG3("Unknown => #", i, ", Channel:", chan_scan);
+          WFM_LOGDEBUG1("Unknown => #", i);
           WFM_HEXLOGDEBUG5_SEPARATOR(":", BSSID_scan[0], BSSID_scan[1], BSSID_scan[2], BSSID_scan[3], BSSID_scan[4], BSSID_scan[5]);
           WFM_LOGDEBUG5("SSID:", ssid_scan.c_str(), ", RSSI:", rssi_scan, ", Secured:", (sec_scan == ENC_TYPE_NONE /*WIFI_AUTH_OPEN*/) ? 'n' : 'y');
         }
@@ -222,7 +218,7 @@ uint8_t WiFiMulti_Generic::run(const uint32_t& connectTimeout)
     {
       WFM_LOGINFO0("[run] Connecting BSSID: ");
       WFM_HEXLOGINFO5_SEPARATOR(":", bestBSSID[0], bestBSSID[1], bestBSSID[2], bestBSSID[3], bestBSSID[4], bestBSSID[5]);
-      WFM_LOGINFO5("SSID: ", bestNetwork.ssid, ", Channel: ", bestChannel, ", Best dB: ", bestNetworkDb);
+      WFM_LOGINFO3("SSID: ", bestNetwork.ssid, ", Best dB: ", bestNetworkDb);
       
       WiFi.begin(bestNetwork.ssid, bestNetwork.passphrase);
       
@@ -277,4 +273,4 @@ uint8_t WiFiMulti_Generic::run(const uint32_t& connectTimeout)
   return status;
 }
 
-#endif    // _NINA_WIFIMULTI_GENERIC_IMPL_H_
+#endif    // _PORTENTAH7_WIFIMULTI_GENERIC_IMPL_H_