+++ src/WiFi101/WiFi101_WiFiMulti_Generic_Impl.h
@@ -1,5 +1,5 @@
 /**************************************************************************************************************************************
-  NINA_WiFiMulti_Generic_Impl.h
+  WiFi101_WiFiMulti_Generic_Impl.h
   For any WiFi shields, such as ESP32, ESP8266, Portenta_H7, WiFiNINA W101, W102, W13x, or custom, such as ESP8266/ESP32-AT, etc
   
   WiFiMulti_Generic is a library to adapt the  ESP32/ESP8266 WiFiMulti feature to other WiFi modules
@@ -34,8 +34,8 @@
 
 #pragma once
 
-#ifndef _NINA_WIFIMULTI_GENERIC_IMPL_H_
-#define _NINA_WIFIMULTI_GENERIC_IMPL_H_ 
+#ifndef _WIFI101_WIFIMULTI_GENERIC_IMPL_H_
+#define _WIFI101_WIFIMULTI_GENERIC_IMPL_H_ 
 
 #include <limits.h>
 #include <string.h>
@@ -277,4 +277,4 @@ uint8_t WiFiMulti_Generic::run(const uint32_t& connectTimeout)
   return status;
 }
 
-#endif    // _NINA_WIFIMULTI_GENERIC_IMPL_H_
+#endif    // _WIFI101_WIFIMULTI_GENERIC_IMPL_H_